+++ src/Contracts/SimpleAuditLog.php
@@ -1,0 +1,6 @@
+<?php
+namespace aliirfaan\LaravelSimpleAuditLog\Contracts;
+
+interface SimpleAuditLog
+{
+}