+++ src/Models/SimpleAuditLog.php
@@ -3,8 +3,9 @@
 namespace aliirfaan\LaravelSimpleAuditLog\Models;
 
 use Illuminate\Database\Eloquent\Model;
+use aliirfaan\LaravelSimpleAuditLog\Contracts\SimpleAuditLog as SimpleAuditLogContract;
 
-class AuditLog extends Model
+class SimpleAuditLog extends Model implements SimpleAuditLogContract
 {
     protected $table = 'lsal_audit_logs';
 
@@ -22,13 +23,14 @@ class AuditLog extends Model
         'al_action_type', 
         'al_event_name', 
         'al_correlation_id',
+        'al_parent_correlation_id',
+        'al_is_success',
+        'al_meta',
+        'al_message',
         'al_previous_value', 
         'al_new_value', 
         'al_request', 
-        'al_response', 
-        'al_custom_field_1', 
-        'al_custom_field_2', 
-        'al_custom_field_3', 
+        'al_response',
         'al_ip_addr', 
         'al_server', 
         'al_version'